Clean up free_symbols/diff
In older versions of Sympy, free_symbols would not contain all the
symbols with respect to which the derivative was not trivially zero.
Specically, indexed symbols would not show up in free_symbols (only the
IndexedBase, and the index). Since this has been fixed, we can simplify
the implementation of the diff method.