Make ranges of IndexedSum a keyword argument
Having ranges as positional arguments to IndexedSum and its subclasses
was a violation of the rule that Operations have only their Operands as
positional arguments. Moreover, IndexedSum must match the interface of
QuantumIndexedSum, which is a QuantumOperation, for which the operands
must be "in the same fundamental set". E.g. the positional arguments of
an KetIndexedSum must be instances of State. Only the `term` fulfills
this, but the ranges do not.

The assumption that the positional arguments of a QuantumOperation are a
QuantumExpression is used in the implementation of
QuantumOperation._simplify_scalar, which was causing ranges to be
dropped when calling that method for an indexed sum.